+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNVVT6J34RYJE7PD2RNP0E14"
+        "01KNXVFNVMZVJNHG5WK0G6SMSH"
     )
     r"""The unique identifier of the data source"""